feedback_agent/requirements.txt
anthropic>=0.40.0
psycopg2-binary>=2.9.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
supabase>=2.0.0
